+++ yorijori/src/main/java/org/dnd4/yorijori/comment/entity/Comment.java
@@ -1,4 +1,4 @@
-package org.dnd4.yorijori.domain.recipe.entity;
+package org.dnd4.yorijori.comment.entity;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -7,11 +7,8 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
-import org.dnd4.yorijori.domain.posts.entity.BaseTimeEntity;
-import org.dnd4.yorijori.domain.user.entity.User;
-import org.hibernate.annotations.ColumnDefault;
-
-import com.sun.istack.NotNull;
+import org.dnd4.yorijori.recipe.entity.Recipe;
+import org.dnd4.yorijori.user.entity.User;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -21,24 +18,19 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @Entity
-public class Recipe extends BaseTimeEntity {
-
+public class Comment {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-
-	@NotNull
-	private String title;
-
-	private int step;
-	private int time;
-
-	@ColumnDefault("0")
-	private int viewCount;
-	private String thumnail;
 
 	@ManyToOne
 	@JoinColumn(name = "user_id", nullable = false)
 	private User user;
 
+	@ManyToOne
+	@JoinColumn(name = "recipe_id", nullable = false)
+	private Recipe recipe;
+
+	private String content;
+	private String imageUrl;
 }
