--- yorijori/src/main/java/org/dnd4/yorijori/controller/MainController.java
@@ -1,14 +1,0 @@
-package org.dnd4.yorijori.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class MainController {
-	
-	@GetMapping("/")
-	public String main() {
-		return "Hello World!";
-	}
-	
-}
