--- yorijori/src/main/java/org/dnd4/yorijori/domain/user/entity/YesOrNo.java
@@ -1,5 +1,0 @@
-package org.dnd4.yorijori.domain.user.entity;
-
-public enum YesOrNo {
-	Y, N
-}
