+++ yorijori/src/main/java/org/dnd4/yorijori/rating/entity/Rating.java
@@ -1,4 +1,4 @@
-package org.dnd4.yorijori.domain.recipe.entity;
+package org.dnd4.yorijori.rating.entity;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -6,6 +6,10 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+
+import org.dnd4.yorijori.domain.posts.entity.BaseTimeEntity;
+import org.dnd4.yorijori.recipe.entity.Recipe;
+import org.dnd4.yorijori.user.entity.User;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,17 +19,18 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @Entity
-public class Step {
-
+public class Rating extends BaseTimeEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@ManyToOne
+	@JoinColumn(name = "user_id", nullable = false)
+	private User user;
+
+	@ManyToOne
 	@JoinColumn(name = "recipe_id", nullable = false)
 	private Recipe recipe;
 
-	private String description;
-	private String imageUrl;
-	private int sequence;
+	private double star;
 }
