+++ yorijori/src/main/java/org/dnd4/yorijori/recipe/entity/Recipe.java
@@ -1,4 +1,4 @@
-package org.dnd4.yorijori.domain.recipe.entity;
+package org.dnd4.yorijori.recipe.entity;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -8,7 +8,10 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
 import org.dnd4.yorijori.domain.posts.entity.BaseTimeEntity;
-import org.dnd4.yorijori.domain.user.entity.User;
+import org.dnd4.yorijori.user.entity.User;
+import org.hibernate.annotations.ColumnDefault;
+
+import com.sun.istack.NotNull;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -18,18 +21,24 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @Entity
-public class Rating extends BaseTimeEntity {
+public class Recipe extends BaseTimeEntity {
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
+
+	@NotNull
+	private String title;
+
+	private int step;
+	private int time;
+
+	@ColumnDefault("0")
+	private int viewCount;
+	private String thumnail;
 
 	@ManyToOne
 	@JoinColumn(name = "user_id", nullable = false)
 	private User user;
 
-	@ManyToOne
-	@JoinColumn(name = "recipe_id", nullable = false)
-	private Recipe recipe;
-
-	private double star;
 }
