+++ yorijori/src/main/java/org/dnd4/yorijori/user/entity/YesOrNo.java
@@ -1,0 +1,5 @@
+package org.dnd4.yorijori.user.entity;
+
+public enum YesOrNo {
+	Y, N
+}
