+++ yorijori/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 
 # DataSource
-spring.datasource.url=jdbc:mysql://13.209.68.130:3306/yorijori?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
-spring.datasource.username=yorijori
+spring.datasource.url=jdbc:mysql://localhost:3306/yorijori?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
+spring.datasource.username=root
 spring.datasource.password=1234
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
